Oppdater tabellvisning ved å fjerne unødvendige props og styling
