+++ src/frontend/Komponenter/Behandling/Inngangsvilkår/Medlemskap/Utenlandsopphold.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { FC } from 'react';
 import { formaterNullableIsoDato } from '../../../../App/utils/formatter';
-import TabellVisning from '../../Vilkårpanel/TabellVisning';
+import TabellVisning from '../../Tabell/TabellVisning';
 import { VilkårInfoIkon } from '../../Vilkårpanel/VilkårInformasjonKomponenter';
 import { IUtenlandsopphold } from './typer';
