--- src/frontend/Komponenter/Behandling/Vilkårpanel/TabellVisning.tsx
@@ -1,56 +1,0 @@
-import * as React from 'react';
-import { Detail, HelpText, Label } from '@navikt/ds-react';
-import { BodyShortSmall } from '../../../Felles/Visningskomponenter/Tekster';
-import styled from 'styled-components';
-import { GridTabell } from './GridTabell';
-import { mapIkon, VilkårInfoIkon } from './VilkårInformasjonKomponenter';
-
-export interface Kolonnedata<T> {
-    ikon?: VilkårInfoIkon;
-    tittel?: string;
-    verdier: T[];
-    kolonner: Kolonner<T>[];
-    ikonVisning?: boolean;
-}
-
-export interface Kolonner<T> {
-    overskrift: string;
-    hjelpetekst?: string;
-    tekstVerdi: (data: T) => React.ReactNode;
-}
-
-const FlexDiv = styled.div`
-    display: flex;
-    gap: 1rem;
-`;
-
-function TabellVisning<T>(props: Kolonnedata<T>): React.ReactElement<Kolonnedata<T>> {
-    const { ikon, tittel, verdier, kolonner, ikonVisning = true } = props;
-    return (
-        <GridTabell kolonner={kolonner.length + 1} ikonVisning={ikonVisning}>
-            {ikon && mapIkon(ikon)}
-            {tittel && (
-                <Label size="small" className="tittel" as="h3">
-                    {tittel}
-                </Label>
-            )}
-            {kolonner.map((kolonne, index) => (
-                <FlexDiv className={index === 0 ? 'førsteDataKolonne' : ''} key={index}>
-                    <Detail>
-                        <strong>{kolonne.overskrift}</strong>
-                    </Detail>
-                    {kolonne.hjelpetekst && <HelpText>{kolonne.hjelpetekst}</HelpText>}
-                </FlexDiv>
-            ))}
-            {verdier.map((item) =>
-                kolonner.map((kolonne, index) => (
-                    <BodyShortSmall className={index === 0 ? 'førsteDataKolonne' : ''} key={index}>
-                        {kolonne.tekstVerdi(item) || ''}
-                    </BodyShortSmall>
-                ))
-            )}
-        </GridTabell>
-    );
-}
-
-export default TabellVisning;
